chore: Atom 파일을 tsx가 아닌 ts 파일로 변경했습니다
